requirements.txt
cupy
pillow
torch
